+++ unstructured/partition/email.py
@@ -167,7 +167,7 @@ def extract_attachment_info(
             for item in cdisp:
                 if item.lower() in ("attachment", "inline"):
                     continue
-                key, value = item.split("=")
+                key, value = item.split("=", 1)
                 key = clean_extra_whitespace(key.replace('"', ""))
                 value = clean_extra_whitespace(value.replace('"', ""))
                 attachment_info[clean_extra_whitespace(key)] = clean_extra_whitespace(